+++ config.go
@@ -15,7 +15,8 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
+	"regexp"
 	"strings"
 	"sync"
 
@@ -76,8 +77,12 @@ func (c CollectorName) GetInstance() (collector, error) {
 		return IPMICollector{}, nil
 	case BMCCollectorName:
 		return BMCCollector{}, nil
+	case BMCWatchdogCollectorName:
+		return BMCWatchdogCollector{}, nil
 	case SELCollectorName:
 		return SELCollector{}, nil
+	case SELEventsCollectorName:
+		return SELEventsCollector{}, nil
 	case DCMICollectorName:
 		return DCMICollector{}, nil
 	case ChassisCollectorName:
@@ -122,8 +127,15 @@ type IPMIConfig struct {
 	CollectorArgs    map[CollectorName][]string `yaml:"default_args"`
 	CustomArgs       map[CollectorName][]string `yaml:"custom_args"`
 
+	SELEvents []*IpmiSELEvent `yaml:"sel_events,omitempty"`
 	// Catches all undefined fields and must be empty after parsing.
 	XXX map[string]interface{} `yaml:",inline"`
+}
+
+type IpmiSELEvent struct {
+	Name     string         `yaml:"name"`
+	RegexRaw string         `yaml:"regex"`
+	Regex    *regexp.Regexp `yaml:"-"`
 }
 
 var defaultConfig = IPMIConfig{
@@ -167,6 +179,9 @@ func (s *IPMIConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
 		if err := c.IsValid(); err != nil {
 			return err
 		}
+	}
+	for _, selEvent := range s.SELEvents {
+		selEvent.Regex = regexp.MustCompile(selEvent.RegexRaw)
 	}
 	return nil
 }
@@ -222,7 +237,7 @@ func (sc *SafeConfig) ReloadConfig(configFile string) error {
 	var err error
 
 	if configFile != "" {
-		config, err = ioutil.ReadFile(configFile)
+		config, err = os.ReadFile(configFile)
 		if err != nil {
 			level.Error(logger).Log("msg", "Error reading config file", "error", err)
 			return err